dedupe dodge_jitter
